requirements.txt
pyrebase
pytz
discord.py
growler
icalendar
techlabreactor
sc2reader
aiohttp